GNS 5890 ADS-B Receiver USB-Stick is the world’s smallest ADS-B Receiver. The GNS 5890 is equipped with a high sensitivity ADS-B receiver. http://www.gns-gmbh.com/index.php?id=238&L=1 Global Navigation Solutions claims to have changed a lot with the advent of their v3 firmware (current is v4) and as such have increased the framerate. Additionally they kept an eye to the so called “doughnut effect”. GNS claims that they are not subject to this issue anymore.(Most other ADS gear is) http://www.blog-ads-b.gns-gmbh.com/?p=29
In November of 2012 is was announced that announce that we are able to use our GNS 5890 ADS-B Dongle together with an Android
http://www.blog-ads-b.gns-gmbh.com/?p=69

A video showing the GNS ADS-B solution running on Android was released
https://www.youtube.com/watch?v=e90QDX52dlw

A GNS 5890 ADS-B USB Android Viewer Public Beta was released here:
http://www.gns-gmbh.com/index.php?id=117%2527&L=1

The original Youtube Video had the comment: "Soure is open if you want to develop your own App do not hesitate to contact GNS GmbH info@GNS-GmbH.com.
We will be happy to support you"

I was told on May 13 2013:
"Unfortunately we had to stop our open source approach for ADSB on Android and instead of this , I provide these sources to directly via email. Please understand that we cannot provide detailed support on it , so I hope you can work with it as it is."

I have maintained and compiled the source code for you to make use of here via the Play Store.